Update settings.py
Added text files list (Currently [] but is updated in the scrape_wikicategory.py as the list of names of text files already processed. The get_pages functions checks if the page names does not exist in the drive before putting its output on the queue. It does that by referring to the settings.text_files list